add(DropsCalendars.vue): base drop calendars
