+++ components/drops/BasicDropCard.vue
@@ -22,18 +22,7 @@
                 class="font-bold overflow-hidden text-ellipsis whitespace-nowrap text-xl"
                 >{{ name }}</span
               >
-              <div
-                v-if="dropPrefix"
-                class="flex items-center rounded-full bg-neutral-3 dark:bg-neutral-11 px-[6px] py-[3px] h-[22px] gap-[0.375rem]">
-                <img
-                  v-if="dropIcon"
-                  :src="dropIcon"
-                  class="w-[12.5px] aspect-square flex-shrink-0"
-                  alt="chain icon" />
-                <span class="text-sm uppercase">{{
-                  UserFriendlyPrefixNameMap[dropPrefix] ?? dropPrefix
-                }}</span>
-              </div>
+              <DropsChainPill :prefix="dropPrefix" />
             </div>
 
             <div v-if="dropCreator" class="flex gap-2 items-center">
@@ -72,14 +61,8 @@
 import type { Prefix } from '@kodadot1/static'
 import { DropStatus } from '@/components/drops/useDrops'
 
-const UserFriendlyPrefixNameMap: Partial<Record<Prefix, string>> = {
-  ahp: 'dot',
-  base: 'base',
-  ahk: 'ahk',
-}
-
 const emit = defineEmits(['click'])
-const props = withDefaults(
+withDefaults(
   defineProps<{
     image: string | undefined
     name: string
@@ -110,9 +93,6 @@ const props = withDefaults(
 )
 
 const { placeholder } = useTheme()
-const { getChainIcon } = useIcon()
-
-const dropIcon = computed(() => getChainIcon(props.dropPrefix))
 </script>
 <style scoped lang="scss">
 @import '@/assets/styles/abstracts/variables';
